UPSTREAM: tracing: Require all trace events to have a TRACE_SYSTEM
The creation of the trace event directory requires that a TRACE_SYSTEM is
defined that the trace event directory is added within the system it was
defined in.

The code handled the case where a TRACE_SYSTEM was not added, and would
then add the event at the events directory. But nothing should be doing
this. This code also prevents the implementation of creating dynamic
dentrys for the eventfs system.

As this path has never been hit on correct code, remove it. If it does get
hit, issues a WARN_ON_ONCE() and return ENODEV.

Link: https://lkml.kernel.org/r/1690568452-46553-2-git-send-email-akaher@vmware.com

BUG=b:360162015
TEST=find /sys/kernel/tracing

Change-Id: Icf94681a8589e391db230f17cf7b185d52b5a231
Signed-off-by: Ajay Kaher <akaher@vmware.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
(cherry picked from commit ee41106a12c76f38d0cf82ef17809fa62757151d)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5806882
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Steven Rostedt <rostedt@google.com>
Commit-Queue: Steven Rostedt <rostedt@google.com>
Reviewed-by: Masami Hiramatsu <mhiramat@google.com>
Auto-Submit: Steven Rostedt <rostedt@google.com>