UPSTREAM: eventfs: Implement removal of meta data from eventfs
When events are removed from tracefs, the eventfs must be aware of this.
The eventfs_remove() removes the meta data from eventfs so that it will no
longer create the files associated with that event.

When an instance is removed from tracefs, eventfs_remove_events_dir() will
remove and clean up the entire "events" directory.

The helper function eventfs_remove_rec() is used to clean up and free the
associated data from eventfs for both of the added functions. SRCU is used
to protect the lists of meta data stored in the eventfs. The eventfs_mutex
is used to protect the content of the items in the list.

As lookups may be happening as deletions of events are made, the freeing
of dentry/inodes and relative information is done after the SRCU grace
period has passed.

Link: https://lkml.kernel.org/r/1690568452-46553-9-git-send-email-akaher@vmware.com

BUG=b:360162015
TEST=find /sys/kernel/tracing

Signed-off-by: Ajay Kaher <akaher@vmware.com>
Co-developed-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Tested-by: Ching-lin Yu <chinglinyu@google.com>
Reported-by: kernel test robot <lkp@intel.com>
Closes: https://lore.kernel.org/oe-kbuild-all/202305030611.Kas747Ev-lkp@intel.com/
Change-Id: I19760e16f35033bf8765ca7569bbbb47966311a4
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
(cherry picked from commit 5bdcd5f5331a276a3ddf1fba8605986d0d15298a)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5791652
Commit-Queue: Steven Rostedt <rostedt@google.com>
Tested-by: Steven Rostedt <rostedt@google.com>
Reviewed-by: Masami Hiramatsu <mhiramat@google.com>
Auto-Submit: Steven Rostedt <rostedt@google.com>
Reviewed-by: Chinglin Yu <chinglinyu@chromium.org>