UPSTREAM: eventfs: Remove extra dget() in eventfs_create_events_dir()
The creation of the top events directory does a dget() at the end of the
creation in eventfs_create_events_dir() with a comment saying the final
dput() will happen when it is removed. The problem is that a dget() is
already done on the dentry when it was created with tracefs_start_creating()!
The dget() now just causes a memory leak of that dentry.

Remove the extra dget() as the final dput() in the deletion of the events
directory actually matches the one in tracefs_start_creating().

Link: https://lore.kernel.org/linux-trace-kernel/20231031124229.4f2e3fa1@gandalf.local.home

BUG=b:360162015
TEST=find /sys/kernel/tracing

Cc: Masami Hiramatsu <mhiramat@kernel.org>
Cc: Mark Rutland <mark.rutland@arm.com>
Fixes: 5790b1fb3d672 ("eventfs: Remove eventfs_file and just use eventfs_inode")
Change-Id: If1ab39170bc06039257f28c1c9ef38b05dce6f29
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
(cherry picked from commit 77bc4d4921bd3497678ba8e7f4e480de35692f05)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5792709
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Masami Hiramatsu <mhiramat@google.com>
Commit-Queue: Steven Rostedt <rostedt@google.com>
Auto-Submit: Steven Rostedt <rostedt@google.com>
Tested-by: Steven Rostedt <rostedt@google.com>