UPSTREAM: eventfs: Remove "is_freed" union with rcu head
The eventfs_inode->is_freed was a union with the rcu_head with the
assumption that when it was on the srcu list the head would contain a
pointer which would make "is_freed" true. But that was a wrong assumption
as the rcu head is a single link list where the last element is NULL.

Instead, split the nr_entries integer so that "is_freed" is one bit and
the nr_entries is the next 31 bits. As there shouldn't be more than 10
(currently there's at most 5 to 7 depending on the config), this should
not be a problem.

Link: https://lkml.kernel.org/r/20231101172649.049758712@goodmis.org

BUG=b:360162015
TEST=find /sys/kernel/tracing

Cc: stable@vger.kernel.org
Cc: Mark Rutland <mark.rutland@arm.com>
Cc: Andrew Morton <akpm@linux-foundation.org>
Cc: Ajay Kaher <akaher@vmware.com>
Fixes: 63940449555e7 ("eventfs: Implement eventfs lookup, read, open functions")
Reviewed-by: Masami Hiramatsu (Google) <mhiramat@kernel.org>
Change-Id: Icc1f14eec27831fa5ee7dd044bc94bd6c479b7a2
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
(cherry picked from commit f2f496370afcbc5227d7002da28c74b91fed12ff)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5792711
Tested-by: Steven Rostedt <rostedt@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Suleiman Souhlal <suleiman@google.com>
Commit-Queue: Steven Rostedt <rostedt@google.com>
Auto-Submit: Steven Rostedt <rostedt@google.com>