UPSTREAM: eventfs: Delete eventfs_inode when the last dentry is freed
There exists a race between holding a reference of an eventfs_inode dentry
and the freeing of the eventfs_inode. If user space has a dentry held long
enough, it may still be able to access the dentry's eventfs_inode after it
has been freed.

To prevent this, have he eventfs_inode freed via the last dput() (or via
RCU if the eventfs_inode does not have a dentry).

This means reintroducing the eventfs_inode del_list field at a temporary
place to put the eventfs_inode. It needs to mark it as freed (via the
list) but also must invalidate the dentry immediately as the return from
eventfs_remove_dir() expects that they are. But the dentry invalidation
must not be called under the eventfs_mutex, so it must be done after the
eventfs_inode is marked as free (put on a deletion list).

Link: https://lkml.kernel.org/r/20231101172650.123479767@goodmis.org

BUG=b:360162015
TEST=find /sys/kernel/tracing

Cc: stable@vger.kernel.org
Cc: Masami Hiramatsu <mhiramat@kernel.org>
Cc: Mark Rutland <mark.rutland@arm.com>
Cc: Andrew Morton <akpm@linux-foundation.org>
Cc: Ajay Kaher <akaher@vmware.com>
Fixes: 5bdcd5f5331a2 ("eventfs: Implement removal of meta data from eventfs")
Change-Id: I900176eacf350ad633e447294f9479837ff24fa9
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
(cherry picked from commit 020010fbfa202aa528a52743eba4ab0da3400a4e)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5792716
Auto-Submit: Steven Rostedt <rostedt@google.com>
Tested-by: Steven Rostedt <rostedt@google.com>
Commit-Queue: Steven Rostedt <rostedt@google.com>
Reviewed-by: Suleiman Souhlal <suleiman@google.com>