UPSTREAM: eventfs: Remove expectation that ei->is_freed means ei->dentry == NULL
The logic to free the eventfs_inode (ei) use to set is_freed and clear the
"dentry" field under the eventfs_mutex. But that changed when a race was
found where the ei->dentry needed to be cleared when the last dput() was
called on it. But there was still logic that checked if ei->dentry was not
NULL and is_freed is set, and would warn if it was.

But since that situation was changed and the ei->dentry isn't cleared
until the last dput() is called on it while the ei->is_freed is set, do
not test for that condition anymore, and change the comments to reflect
that.

Link: https://lkml.kernel.org/r/20231120235154.265826243@goodmis.org

BUG=b:360162015
TEST=find /sys/kernel/tracing

Cc: Masami Hiramatsu <mhiramat@kernel.org>
Cc: Andrew Morton <akpm@linux-foundation.org>
Fixes: 020010fbfa20 ("eventfs: Delete eventfs_inode when the last dentry is freed")
Reported-by: Mark Rutland <mark.rutland@arm.com>
Change-Id: Ic1d682eabb7e41609b0eaaac4798a602c7ea0e7a
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
(cherry picked from commit 88903daecacf03b1e5636e1b5f18bda5b07030fc)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5792719
Auto-Submit: Steven Rostedt <rostedt@google.com>
Reviewed-by: Suleiman Souhlal <suleiman@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Steven Rostedt <rostedt@google.com>
Tested-by: Steven Rostedt <rostedt@google.com>