UPSTREAM: eventfs: Fix events beyond NAME_MAX blocking tasks
Eventfs uses simple_lookup(), however, it will fail if the name of the
entry is beyond NAME_MAX length. When this error is encountered, eventfs
still tries to create dentries instead of skipping the dentry creation.
When the dentry is attempted to be created in this state d_wait_lookup()
will loop forever, waiting for the lookup to be removed.

Fix eventfs to return the error in simple_lookup() back to the caller
instead of continuing to try to create the dentry.

Link: https://lore.kernel.org/linux-trace-kernel/20231210213534.497-1-beaub@linux.microsoft.com

BUG=b:360162015
TEST=find /sys/kernel/tracing

Fixes: 63940449555e ("eventfs: Implement eventfs lookup, read, open functions")
Link: https://lore.kernel.org/linux-trace-kernel/20231208183601.GA46-beaub@linux.microsoft.com/
Change-Id: Ic5639c40610f52526baa3554e5d96b6a8ff5d517
Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
(cherry picked from commit 5eaf7f0589c0d88178f0fbeebe0e0b7108258707)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5792725
Auto-Submit: Steven Rostedt <rostedt@google.com>
Reviewed-by: Suleiman Souhlal <suleiman@google.com>
Commit-Queue: Steven Rostedt <rostedt@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Steven Rostedt <rostedt@google.com>