UPSTREAM: eventfs: Fix bitwise fields for "is_events"
A flag was needed to denote which eventfs_inode was the "events"
directory, so a bit was taken from the "nr_entries" field, as there's not
that many entries, and 2^30 is plenty. But the bit number for nr_entries
was not updated to reflect the bit taken from it, which would add an
unnecessary integer to the structure.

Link: https://lore.kernel.org/linux-trace-kernel/20240102151832.7ca87275@gandalf.local.home

BUG=b:360162015
TEST=find /sys/kernel/tracing

Cc: stable@vger.kernel.org
Cc: Masami Hiramatsu <mhiramat@kernel.org>
Cc: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Fixes: 7e8358edf503e ("eventfs: Fix file and directory uid and gid ownership")
Change-Id: I5dd09283283f9de599f636c4d507e455127f57b1
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
(cherry picked from commit fd56cd5f6d76e93356d9520cf9dabffe1e3d1aa0)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5792729
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Steven Rostedt <rostedt@google.com>
Tested-by: Steven Rostedt <rostedt@google.com>
Reviewed-by: Suleiman Souhlal <suleiman@google.com>
Auto-Submit: Steven Rostedt <rostedt@google.com>